requirements.txt
aiogram>=3.0.0
llama-cpp-python